Send articles to log via API
